DNN Model moved to a separate module
